Correctly decide whether to mallsell or autosell
The script currently uses the wellStocked() function to decide
whether to autosell or mall sell an item but to do that it needs to know
the lowest allowed price in the mall which should be
max(100, autosellPrice(item) * 2) but it previously calculated the min of
these values instead. This commit changes it to max as it should be.